requirements.txt
edge-tts>=6.1.0
moviepy>=1.0.3
Pillow>=10.0.0
numpy
youtube-transcript-api==0.6.3
requests==2.32.3
python-dotenv==1.0.1
google-api-python-client==2.147.0
google-auth==2.35.0
